// src/app/admin-dashboard/models/req-approved.model.ts
export class RequestApproved {
  '_id': string;
  'hostFees': number;
  'asicName': string;
  'asicStatus': boolean;
  'expired': boolean;
  'totalMined': number;
  'startDate': Date;
  'userID': string;
  'asicID': string;
  'createdAt': Date;
  'updatedAt': Date;
  'address': string;
  'workerID': string;
}